Removing activity workflow write to s3
