+++ backend/api/model.py
@@ -1,15 +1,13 @@
 from concurrent import futures
-from datetime import date, datetime
+from datetime import datetime
 import json
 import os
 from typing import Tuple, Dict, List, Callable, Any
 from zipfile import ZipFile
 from io import BytesIO
 from collections import defaultdict
-import pandas as pd
 
 from api.models import (
-    github_activity,
     install_activity,
     plugin as plugin_model,
     plugin_blocked,
@@ -17,7 +15,7 @@ from api.models import (
 )
 from utils.github import get_github_metadata, get_artifact
 from utils.pypi import query_pypi, get_plugin_pypi_metadata
-from api.s3 import get_cache, cache, write_data
+from api.s3 import get_cache, cache
 from utils.utils import (
     render_description,
     send_alert,
@@ -28,8 +26,6 @@ from utils.utils import (
 from utils.datadog import report_metrics
 from api.zulip import notify_new_packages
 import boto3
-import snowflake.connector as sc
-from dateutil.relativedelta import relativedelta
 import logging
 
 LOGGER = logging.getLogger()
@@ -433,264 +429,3 @@ def get_categories_mapping(version: str) -> Dict[str, List]:
     """
     mappings = get_cache(f'category/{version.replace(":", "/")}.json')
     return mappings or {}
-
-
-def _execute_query(query, schema):
-    SNOWFLAKE_USER = os.getenv('SNOWFLAKE_USER')
-    SNOWFLAKE_PASSWORD = os.getenv('SNOWFLAKE_PASSWORD')
-    ctx = sc.connect(
-        user=SNOWFLAKE_USER,
-        password=SNOWFLAKE_PASSWORD,
-        account="CZI-IMAGING",
-        warehouse="IMAGING",
-        database="IMAGING",
-        schema=schema
-    )
-    return ctx.execute_string(query)
-
-
-def update_activity_data():
-    LOGGER.info("Starting data refresh for metrics")
-    _update_activity_timeline_data()
-    _update_recent_activity_data()
-    repo_to_plugin_dict = _get_repo_to_plugin_dict()
-    _update_latest_commits(repo_to_plugin_dict)
-    _update_commit_activity(repo_to_plugin_dict)
-    LOGGER.info("Completed data refresh for metrics successfully")
-
-
-def _update_activity_timeline_data():
-    """
-    Update existing caches to reflect new activity data.
-    """
-    query = """
-        SELECT 
-            LOWER(file_project) AS name, DATE_TRUNC('month', timestamp) AS month, COUNT(*) AS num_downloads
-        FROM
-            imaging.pypi.labeled_downloads
-        WHERE 
-            download_type = 'pip'
-            AND project_type = 'plugin'
-        GROUP BY name, month
-        ORDER BY name, month
-        """
-    cursor_list = _execute_query(query, "PYPI")
-    csv_string = "PROJECT,MONTH,NUM_DOWNLOADS_BY_MONTH\n"
-    for cursor in cursor_list:
-        for row in cursor:
-            csv_string += str(row[0]) + ',' + str(row[1]) + ',' + str(row[2]) + '\n'
-    write_data(csv_string, "activity_dashboard_data/plugin_installs.csv")
-
-
-def _process_for_dates(limit):
-    end_date = date.today().replace(day=1) + relativedelta(months=-1)
-    start_date = end_date + relativedelta(months=-limit + 1)
-    dates = pd.date_range(start=start_date, periods=limit, freq='MS')
-    return start_date, end_date, dates
-
-
-def _process_usage_timeline(plugin_df, limit):
-    date_format = '%Y-%m-%d'
-    start_date, end_date, dates = _process_for_dates(limit)
-    plugin_df = plugin_df[(plugin_df['MONTH'] >= start_date.strftime(date_format)) & (
-                plugin_df['MONTH'] <= end_date.strftime(date_format))]
-    result = []
-    for cur_date in dates:
-        if cur_date in plugin_df['MONTH'].values:
-            row = plugin_df[plugin_df['MONTH'] == cur_date]
-            installs = int(str(row.NUM_DOWNLOADS_BY_MONTH).split()[1])
-        else:
-            installs = 0
-        result.append({'timestamp': int(cur_date.timestamp()) * 1000, 'installs': installs})
-    return result
-
-
-def _process_maintenance_timeline(commit_activity, limit):
-    start_date, end_date, dates = _process_for_dates(limit)
-    maintenance_dict = {}
-    for commit_obj in commit_activity:
-        commit_obj_datetime = datetime.utcfromtimestamp(commit_obj['timestamp'] / 1000)
-        if start_date <= commit_obj_datetime.date() <= end_date:
-            maintenance_dict[commit_obj_datetime] = commit_obj
-    result = []
-    for cur_date in dates:
-        if cur_date in maintenance_dict:
-            result.append(maintenance_dict[cur_date])
-        else:
-            result.append({'timestamp': int(cur_date.timestamp()) * 1000, 'commits': 0})
-    return result
-
-
-def _process_for_stats(plugin_df):
-    if len(plugin_df) == 0:
-        return {}
-
-    return {'totalInstalls': int(plugin_df['NUM_DOWNLOADS_BY_MONTH'].sum())}
-
-
-def _update_recent_activity_data(number_of_time_periods=30, time_granularity='DAY'):
-    """
-    Update existing caches to reflect recent activity data.
-    """
-    query = f"""
-        SELECT 
-            LOWER(file_project) AS name, COUNT(*) AS num_downloads
-        FROM
-            imaging.pypi.labeled_downloads
-        WHERE 
-            download_type = 'pip'
-            AND project_type = 'plugin'
-            AND timestamp > DATEADD({time_granularity}, {number_of_time_periods * -1}, CURRENT_DATE)
-        GROUP BY name     
-        ORDER BY name
-    """
-    cursor_list = _execute_query(query, "PYPI")
-    data = {}
-    for cursor in cursor_list:
-        for row in cursor:
-            data[row[0]] = row[1]
-
-    write_data(json.dumps(data), "activity_dashboard_data/recent_installs.json")
-
-
-def _update_repo_to_plugin_dict(repo_to_plugin_dict: dict, plugin_obj: dict):
-    code_repository = plugin_obj.get('code_repository')
-    if code_repository:
-        repo_to_plugin_dict[code_repository.replace('https://github.com/', '')] = plugin_obj['name'].lower()
-    return repo_to_plugin_dict
-
-
-def _get_repo_to_plugin_dict():
-    index_json = get_index()
-    hidden_plugins = get_hidden_plugins()
-    excluded_plugins = get_excluded_plugins()
-    repo_to_plugin_dict = {}
-    for public_plugin_obj in index_json:
-        repo_to_plugin_dict = _update_repo_to_plugin_dict(repo_to_plugin_dict, public_plugin_obj)
-    for excluded_plugin_name, excluded_plugin_visibility in excluded_plugins.items():
-        if excluded_plugin_visibility == "hidden":
-            excluded_plugin_obj = get_plugin(excluded_plugin_name, hidden_plugins[excluded_plugin_name])
-        else:
-            excluded_plugin_obj = get_plugin(excluded_plugin_name, None)
-        repo_to_plugin_dict = _update_repo_to_plugin_dict(repo_to_plugin_dict, excluded_plugin_obj)
-    return repo_to_plugin_dict
-
-
-def _get_repo_from_plugin(plugin):
-    plugin_metadata = get_plugin(plugin)
-    if plugin_metadata:
-        repo_url = plugin_metadata.get('code_repository')
-        if repo_url:
-            return repo_url.replace('https://github.com/', '')
-    return None
-
-
-def _update_latest_commits(repo_to_plugin_dict):
-    """
-    Get the latest commit occurred for the plugin
-    """
-    query = f"""
-        SELECT 
-            repo AS name, MAX(commit_author_date) AS latest_commit
-        FROM 
-            imaging.github.commits
-        WHERE 
-            repo_type = 'plugin'
-        GROUP BY name
-        ORDER BY name
-    """
-    cursor_list = _execute_query(query, "GITHUB")
-    data = {}
-    for cursor in cursor_list:
-        for row in cursor:
-            repo = row[0]
-            if repo in repo_to_plugin_dict:
-                plugin = repo_to_plugin_dict[repo]
-                timestamp = int(pd.to_datetime(row[1]).strftime("%s")) * 1000
-                data[plugin] = timestamp
-    write_data(json.dumps(data), "activity_dashboard_data/latest_commits.json")
-
-
-def _update_commit_activity(repo_to_plugin_dict):
-    """
-    Get the commit activity occurred for the plugin in the past year
-    """
-    query = f"""
-        SELECT 
-            repo AS name, DATE_TRUNC('month', TO_DATE(commit_author_date)) AS month, COUNT(*) AS commit_count
-        FROM 
-            imaging.github.commits
-        WHERE 
-            repo_type = 'plugin'
-        GROUP BY name, month
-        ORDER BY name, month
-    """
-    cursor_list = _execute_query(query, "GITHUB")
-    data = {}
-    for cursor in cursor_list:
-        for repo, month, commit_count in cursor:
-            if repo in repo_to_plugin_dict:
-                plugin = repo_to_plugin_dict[repo]
-                timestamp = int(pd.to_datetime(month).strftime("%s")) * 1000
-                commits = int(commit_count)
-                data.setdefault(plugin, []).append({'timestamp': timestamp, 'commits': commits})
-    for plugin in data:
-        data[plugin] = sorted(data[plugin], key=lambda x: (x['timestamp']))
-    write_data(json.dumps(data), "activity_dashboard_data/commit_activity.json")
-
-
-def _get_usage_data(plugin: str, limit: int) -> Dict[str, Any]:
-    """
-    Fetches plugin usage_data from s3 or dynamo based on the in_test variable
-    :returns (dict[str, Any]): A dict with the structure
-    {'timeline': List, 'stats': dict[str, int]}
-    :params str plugin: Name of the plugin in lowercase.
-    :params int limit: Sets the number of records to be fetched for timeline.
-    """
-    return {
-        "timeline": install_activity.get_timeline(plugin, limit) if limit else [],
-        "stats": {
-            "total_installs": install_activity.get_total_installs(plugin),
-            "installs_in_last_30_days": install_activity.get_recent_installs(plugin, 30)
-        },
-    }
-
-
-def _get_maintenance_data(plugin: str, repo: Any, limit: int) -> Dict[str, Any]:
-    """
-    Fetches plugin maintenance_data from s3 or dynamo based on the in_test variable
-    :returns (dict[str, Any]): A dict with the structure {'timeline': List, 'stats': Dict[str, int]}
-    :params str plugin: Name of the plugin in lowercase.
-    :params repo: Parameter used if use_dynamo_for_maintenance is true
-    :params int limit: Sets the number of records to be fetched for timeline.
-    """
-    return {
-        "timeline": github_activity.get_timeline(plugin, repo, limit) if limit else [],
-        "stats": {
-            "total_commits": github_activity.get_total_commits(plugin, repo),
-            "latest_commit_timestamp": github_activity.get_latest_commit(plugin, repo),
-        },
-    }
-
-
-def get_metrics_for_plugin(plugin: str, limit: str,) -> Dict[str, Any]:
-    """
-    Fetches plugin metrics from s3 or dynamo based on the in_test variable
-    :return dict[str, Any]: A map with entries for usage and maintenance
-
-    :params str plugin: Name of the plugin in lowercase for which usage data needs to be
-    fetched.
-    :params str limit_str: Number of records to be fetched for timeline. Defaults to 0
-    for invalid number.
-    """
-    repo = _get_repo_from_plugin(plugin)
-    plugin = plugin.lower()
-    month_delta = 0
-
-    if limit.isdigit() and limit != "0":
-        month_delta = max(int(limit), 0)
-
-    return {
-        "usage": _get_usage_data(plugin, month_delta),
-        "maintenance": _get_maintenance_data(plugin, repo, month_delta),
-    }
